split result into a file
