+++ client/client.go
@@ -4,8 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
-	"io"
 	"net/http"
 
 	"github.com/go-jarvis/rum-gonic/pkg/operator"
@@ -59,26 +57,4 @@ func newRequest(ctx context.Context, api operator.APIOperator, meta Meta) (*http
 	req.Header = http.Header(meta)
 
 	return req, nil
-}
-
-type Result struct {
-	response *http.Response
-}
-
-func (r *Result) Bind(receiver interface{}) (Meta, error) {
-	meta := Meta(r.response.Header)
-	if receiver == nil {
-		return meta, nil
-	}
-
-	b, err := io.ReadAll(r.response.Body)
-
-	if err != nil {
-		return meta, err
-	}
-	defer r.response.Body.Close()
-
-	fmt.Println(string(b))
-
-	return meta, json.Unmarshal(b, receiver)
 }
